+++ src/engine/location.rs
@@ -32,11 +32,11 @@ impl PartialEq for Real2D {
     }
 }
 
-/// A structure describing a two-dimensional, i64 position, for use in discrete fields such as a grid.
+/// A structure describing a two-dimensional, i32 position, for use in discrete fields such as a grid.
 #[derive(Clone, Hash, Copy, Debug)]
 pub struct Int2D {
-    pub x: i64,
-    pub y: i64,
+    pub x: i32,
+    pub y: i32,
 }
 
 impl fmt::Display for Int2D {
