+++ src/engine/priority.rs
@@ -5,11 +5,11 @@ use std::fmt;
 #[derive(Clone, Debug)]
 pub struct Priority {
     pub time: f32,
-    pub ordering: i64,
+    pub ordering: i32,
 }
 
 impl Priority {
-    pub fn new(the_time: f32, the_ordering: i64) -> Priority {
+    pub fn new(the_time: f32, the_ordering: i32) -> Priority {
         Priority {
             time: the_time,
             ordering: the_ordering,
